requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.57.2
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.1
SQLAlchemy==2.0.36
Jinja2==3.1.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.1
starlette==0.27.0
anyio==3.7.1
httpx==0.28.1
typing_extensions==4.12.2
soundfile==0.12.1
numpy<2
websockets==12.0